App: support auto loading PropertyXLink'ed documents
PropertyXLink can now ask App::Application::openDocument to recursively
load any external document referenced

+++ src/App/Document.cpp
@@ -1539,7 +1539,7 @@ bool Document::save (void)
 }
 
 // Open the document
-void Document::restore (void)
+void Document::restore (bool delaySignal)
 {
     // clean up if the document is not empty
     // !TODO mind exeptions while restoring!
@@ -1588,14 +1588,33 @@ void Document::restore (void)
     signalRestoreDocument(reader);
     reader.readFiles(zipstream);
 
+    if(!delaySignal)
+        afterRestore();
+}
+
+void Document::afterRestore(bool checkXLink) {
     // reset all touched
     for (std::map<std::string,DocumentObject*>::iterator It= d->objectMap.begin();It!=d->objectMap.end();++It) {
         It->second->connectRelabelSignals();
         It->second->onDocumentRestored();
         It->second->ExpressionEngine.onDocumentRestored();
-        It->second->purgeTouched();
-    }
 
+        bool purge = true;
+        if(checkXLink) {
+            std::vector<Property*> props;
+            It->second->getPropertyList(props);
+            for(auto prop : props) {
+                if(prop->isDerivedFrom(PropertyXLink::getClassTypeId()) &&
+                !static_cast<PropertyXLink*>(prop)->isRestored())
+                {
+                    purge = false;
+                    break;
+                }
+            }
+        }
+        if(purge)
+            It->second->purgeTouched();
+    }
     GetApplication().signalFinishRestoreDocument(*this);
 }
